GrpcMaxCallRecvMsgSizeFlag: int -> uint64 flag

+++ cmd/flags.go
@@ -214,7 +214,7 @@ var (
 		Usage: "The path to a YAML file with chain config values",
 	}
 	// GrpcMaxCallRecvMsgSizeFlag defines the max call message size for GRPC
-	GrpcMaxCallRecvMsgSizeFlag = &cli.IntFlag{
+	GrpcMaxCallRecvMsgSizeFlag = &cli.Uint64Flag{
 		Name:  "grpc-max-msg-size",
 		Usage: "Integer to define max recieve message call size (default: 4194304 (for 4MB))",
 		Value: 1 << 22,
